requirements.txt
arrow==1.1.1
asgiref==3.4.1
attrs==21.2.0
base58==0.2.2
bitcoin==1.1.39
blockcypher==1.0.92
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==6.6
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-cors-headers==3.8.0
django-formtools==2.3
django-otp==1.1.1
django-phonenumber-field==5.2.0
django-two-factor-auth==1.13.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
docker-py==1.8.0
ecdsa==0.17.0
flake8==3.9.2
funcsigs==1.0.2
future==0.18.2
idna==3.2
iniconfig==1.1.1
jsonrpcclient==2.0.1
jsonrpcserver==3.1.1
jsonschema==3.2.0
mccabe==0.6.1
mnemonic==0.13
oauthlib==3.1.1
packaging==21.0
path==16.2.0
path.py==12.5.0
pbkdf2==1.3
pexpect==4.8.0
phonenumbers==8.12.31
phonenumberslite==8.12.31
Pillow==8.3.2
pluggy==1.0.0
protobuf==3.0.0a3
ptyprocess==0.7.0
py==1.10.0
pyaes==1.6.1
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.4
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
pywallet==0.1.0
PyYAML==5.4.1
qrcode==6.1
requests==2.11.1
requests-oauthlib==1.3.0
sha256==0.1
six==1.16.0
sqlparse==0.4.2
tabulate==0.8.9
toml==0.10.2
two1==3.10.9
urllib3==1.26.6
websocket-client==1.2.1
